+++ CHANGELOG.md
@@ -6,3 +6,10 @@ All changes in this project will be recorded in this file
 -  Support for **JPG** images
 #### Security
 -  In some large images the *libjpeg* release may cause failures (failed to create temporary file). It is solved if you run your program in administrator mode (pero es muy soso)
+
+### [0.1.0] - 2020-12-20
+#### Added
+-  Support for **PNG** images
+-  Now you can't deliberately resize images
+#### Changed
+-  Changed **struct image_t** to **class Image**